Move fabric/neoforge specific classes to their own package
