+++ common/src/main/java/xyz/eclipseisoffline/customtimecycle/mixin/ServerClockManagerMixin.java
@@ -8,7 +8,6 @@ import net.minecraft.server.MinecraftServer;
 import net.minecraft.world.clock.ClockManager;
 import net.minecraft.world.clock.ClockNetworkState;
 import net.minecraft.world.clock.ClockTimeMarker;
-import net.minecraft.world.clock.ClockTimeMarkers;
 import net.minecraft.world.clock.ServerClockManager;
 import net.minecraft.world.clock.WorldClock;
 import net.minecraft.world.level.saveddata.SavedData;
