+++ fabric/src/main/java/xyz/eclipseisoffline/customtimecycle/fabric/client/CustomTimeCycleModMenuCompat.java
@@ -1,8 +1,9 @@
-package xyz.eclipseisoffline.customtimecycle.client;
+package xyz.eclipseisoffline.customtimecycle.fabric.client;
 
 import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
 import xyz.eclipseisoffline.customtimecycle.CustomTimeCycle;
+import xyz.eclipseisoffline.customtimecycle.client.ManageTimeCycleScreen;
 
 public class CustomTimeCycleModMenuCompat implements ModMenuApi {
