+++ neoforge/src/main/resources/META-INF/neoforge.mods.toml
@@ -13,9 +13,6 @@ description='''${mod_description}'''
 [[mixins]]
 config="customtimecycle.mixins.json"
 
-[[mixins]]
-config="customtimecycle-neoforge.mixins.json"
-
 [[dependencies]]
     modId="neoforge"
     type="required"
