--- neoforge/src/main/resources/customtimecycle-neoforge.mixins.json
@@ -1,11 +1,0 @@
-{
-  "required": true,
-  "minVersion": "0.8",
-  "package": "xyz.eclipseisoffline.customtimecycle.mixin",
-  "compatibilityLevel": "JAVA_25",
-  "mixins": [
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
